+++ lua/github-theme/terminal/alacritty.lua
@@ -1,5 +1,5 @@
-local util = require('github-theme.util')
 local palette = require('github-theme.palette')
+local util = require('github-theme.util')
 
 ---Generate github theme for Alacritty terminal.
 ---@param cfg gt.ConfigSchema