+++ lua/github-theme/terminal/tmux.lua
@@ -1,5 +1,5 @@
-local util = require('github-theme.util')
 local palette = require('github-theme.palette')
+local util = require('github-theme.util')
 
 ---Generate github theme for tmux.
 ---@param cfg gt.ConfigSchema